requirements.txt
# Core Streamlit
streamlit>=1.28.0

# LlamaIndex for orchestration
llama-index>=0.9.0
llama-index-core>=0.9.0

# LLM providers
friendli>=1.0.0
boto3>=1.28.0  # For Bedrock fallback

# Vector store
weaviate-client>=3.24.0

# AWS SDK
boto3>=1.28.0
botocore>=1.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0
altair>=5.1.0

# CloudFormation
cfn-flip>=1.3.0
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
